sentiment/requirements.txt
nltk
viam-sdk